+++ sdk/keyvault/keyvault-admin/recordings/node/aborting_keyvaultaccesscontrolclients_requests/recording_can_abort_listroleassignments.js
@@ -1,5 +1,5 @@
 let nock = require('nock');
 
-module.exports.hash = "014b26d2a6855e2963a4fd04699658c3";
+module.exports.hash = "65288b97bc515cb8ce551ed91f5ccf09";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}